rename login.js -> auth.js ; drop logout route
