+++ src/auth.js
@@ -1,15 +1,19 @@
 // Some insecure authentication functions
-// Check if the user is logged in (... by existence of a cookie!)
+
+import { logErr } from './logger';
+
+// Insecurely check if the user is authenticated.
 export function isLoggedIn(req) {
-  return !!req.cookies?.user; // Insecure: this simply checks if _any_ cookie named 'user' exists, without validation.
+  return !!validateToken(req); // Insecure: this simply checks if the token can be decoded to an object, without any validation.
 }
 
-// Check if the user is an admin (by a cookie attribute!)
+// Check if the user is an admin by the authentication token.
+// This token is not signed, and so vulnerable to manipulation and forgery.
 export function isAdmin(req) {
-  const cookie = req.cookies?.user;
-  if (!cookie) return false;
-  const role = JSON.parse(cookie).role;
-  return role === 'admin'; // Insecure: simply checks the cookie for 'role' parameter. which can be modified by the user.
+  const user = parseToken(req);
+  if (!user) return false;
+  const role = JSON.parse(user)?.role;
+  return role === 'admin'; // Insecure: simply checks the token for 'role' parameter. which can be modified by the user.
 }
 
 // Check if the logged in user's ID matches the userId route parameter
@@ -28,17 +32,24 @@ export function generateToken(user) {
   return Buffer.from(JSON.stringify(user)).toString('base64');
 }
 
-// Really insecure token validation
-// This currently just decodes the token into a user.
-export function validateToken(token) {
+// Really insecure token validation. Basically just checks the token can be decoded into an object.
+export const validateToken = (req) => {
+  const token = parseToken(req);
+  return !!token;
+};
+
+// Extract and parse token into User object from request headers.
+export const parseToken = (req) => {
   try {
+    const token = req.headers?.authorization?.replace('Bearer ', '') || '';
+    if (!token) {
+      return;
+    }
+    // Decode base64 and parse JSON to object
     const decoded = Buffer.from(token, 'base64').toString('ascii');
-    console.log('=== debug: decoded: ' + decoded);
-    const parsedUser = JSON.parse(decoded);
-    console.log('=== debug: parsedUser.email: ' + parsedUser?.email);
-    return parsedUser;
+    return JSON.parse(decoded);
   } catch (err) {
-    console.log('decode token failed:', err.message);
+    logErr('decode token failed', err);
     return;
   }
-}
+};
